Before PostgreSQL

+++ backend/api/views.py
@@ -5,9 +5,9 @@ from rest_framework.permissions import IsAuthenticated, AllowAny
 from recipes.models import Recipe, Tag, Follow, Ingredient, IngredientIndividual, ShoppingList, Favorites
 from rest_framework.response import Response
 from rest_framework.pagination import PageNumberPagination
-from .permissions import OwnerOrReadOnly
+from .permissions import OwnerOrReadOnly, OwnerOnly
 from .serializers import (
-    FollowSerializer, RecipeCreateSerializer, RecipeGetSerializer, TagSerializer, 
+    RecipeCreateSerializer, RecipeGetSerializer, TagSerializer, 
     FollowCreateSerializer, ShoppingListSerializer, FavoritesSerializer, 
     IngredientSerilizer
 )
@@ -21,7 +21,7 @@ from djoser.views import UserViewSet
 from users.serializers import CustomUserSerializer
 from django.contrib.auth import get_user_model
 from rest_framework.generics import get_object_or_404, GenericAPIView, DestroyAPIView
-from rest_framework.status import HTTP_200_OK, HTTP_400_BAD_REQUEST, HTTP_204_NO_CONTENT
+from rest_framework.status import HTTP_200_OK, HTTP_400_BAD_REQUEST, HTTP_204_NO_CONTENT, HTTP_201_CREATED
 from rest_framework.decorators import action
 
 
@@ -31,7 +31,12 @@ User = get_user_model()
 class CustomUserViewSet(UserViewSet):
     queryset = User.objects.all()
     serializer_class = CustomUserSerializer
-    pagination_class = PageNumberPagination
+    pagination_class = LimitOffsetPagination
+    
+    def get_permissions(self):
+        if self.action == 'me' and self.request.user.is_authenticated == False:
+            return (OwnerOnly(), )
+        return super().get_permissions()
 
     @action(
         methods=[
@@ -40,23 +45,20 @@ class CustomUserViewSet(UserViewSet):
         permission_classes=(
             IsAuthenticated,
         ),
-        url_path='subscribe',
+        # url_path='subscribe',
         detail=True,
     )
     def subscribe(self, request, id=None):
-        # user = request.user
         following = get_object_or_404(User, pk=id)
-        serializer = FollowCreateSerializer(data={"user": request.user.id, "following": id},
-                                            context={'request': request})
-        print(serializer)
+        serializer = FollowCreateSerializer(data={"user": request.user, "following": following},
+                                            context={'request': request, 'pk': id})
         serializer.is_valid(raise_exception=True)
         serializer.save()
-        serializer_user = CustomUserSerializer(following, context={'request': request})
-        print(serializer_user)
-        return Response(serializer_user.data)
-    
+        return Response(status=HTTP_201_CREATED, data=serializer.data)
+
     @subscribe.mapping.delete
     def subscribe_delete(self, request, id=None):
+        get_object_or_404(User, pk=id)
         if Follow.objects.filter(user=request.user.id, following=id).exists():
             instance = Follow.objects.get(user=request.user.id, following=id)
             instance.delete()
@@ -70,7 +72,7 @@ class CustomUserViewSet(UserViewSet):
         permission_classes=(
             IsAuthenticated,
         ),
-        # url_path='subscriptions',
+        url_path='subscriptions',
         detail=False,
     )
     def subscriptions(self, request):
@@ -78,11 +80,11 @@ class CustomUserViewSet(UserViewSet):
             user=self.request.user)
         page = self.paginate_queryset(following)
         if page is not None:
-            serializer = FollowSerializer(page, many=True, context={'request': request})
+            serializer = FollowCreateSerializer(page, many=True, context={'request': request})
             return self.get_paginated_response(serializer.data)
 
         print(following)
-        serializer = FollowSerializer(following, many=True, context={'request': request})
+        serializer = FollowCreateSerializer(following, many=True, context={'request': request})
         print(serializer.data)
         return Response(
             serializer.data,
@@ -118,23 +120,26 @@ class RecipeViewSet(viewsets.ModelViewSet):
         detail=True,
     )
     def shopping_cart(self, request, pk=None):
-        recipe = get_object_or_404(
-            Recipe,
-            pk=pk)
-        serializer = ShoppingListSerializer(data={"user": request.user, "recipe": recipe},
-                                            context={'request': request, 'pk': pk})
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data) 
+        if Recipe.objects.filter(pk=pk).exists():
+            recipe = Recipe.objects.get(pk=pk)
+            serializer = ShoppingListSerializer(data={"user": request.user, "recipe": recipe},
+                                                context={'request': request, 'pk': pk})
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return Response(status=HTTP_201_CREATED, data=serializer.data)
+        return Response(status=HTTP_400_BAD_REQUEST, data={"errors": "Такого рецепта не существует"})
 
     @shopping_cart.mapping.delete
     def shopping_cart_delete(self, request, pk=None):
         recipe = get_object_or_404(
             Recipe,
             pk=pk)
-        instance = ShoppingList.objects.get(recipe=recipe, user=request.user)
-        instance.delete()
-        return Response(status=HTTP_204_NO_CONTENT) 
+        if ShoppingList.objects.filter(recipe=recipe, user=request.user).exists():
+            instance = ShoppingList.objects.get(recipe=recipe, user=request.user)
+            instance.delete()
+            return Response(status=HTTP_204_NO_CONTENT)
+        return Response(status=HTTP_400_BAD_REQUEST, data={"errors": "Такого рецепта нет в списке покупок"})
+        
     
     @action(
         methods=[
@@ -147,23 +152,27 @@ class RecipeViewSet(viewsets.ModelViewSet):
         detail=True,
     )
     def favorite(self, request, pk=None):
-        recipe = get_object_or_404(
-            Recipe,
-            pk=pk)
-        serializer = FavoritesSerializer(data={"user": request.user, "recipe": recipe}, 
-                                         context={'request': request, 'pk': pk})
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data) 
-
+        if Recipe.objects.filter(pk=pk).exists():
+            recipe = Recipe.objects.get(pk=pk)
+            serializer = FavoritesSerializer(data={"user": request.user, "recipe": recipe}, 
+                                            context={'request': request, 'pk': pk})
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return Response(status=HTTP_201_CREATED, data=serializer.data)
+        else:
+            return Response(status=HTTP_400_BAD_REQUEST, data={"errors": "Такого рецепта не существует"})
+        
     @favorite.mapping.delete
     def favorite_delete(self, request, pk=None):
         recipe = get_object_or_404(
             Recipe,
             pk=pk)
-        instance = Favorites.objects.get(recipe=recipe, user=request.user)
-        instance.delete()
-        return Response(status=HTTP_204_NO_CONTENT) 
+        if Favorites.objects.filter(recipe=recipe, user=request.user).exists():
+            # recipe = Recipe.objects.get(pk=pk)
+            instance = Favorites.objects.get(recipe=recipe, user=request.user)
+            instance.delete()
+            return Response(status=HTTP_204_NO_CONTENT)
+        return Response(status=HTTP_400_BAD_REQUEST, data={"errors": "Такого рецепта нет в избранном"})
     
     @action(
         methods=[
@@ -203,17 +212,17 @@ class RecipeViewSet(viewsets.ModelViewSet):
     
     
 
-
 class TagViewSet(viewsets.ReadOnlyModelViewSet):
     queryset = Tag.objects.all()
     serializer_class = TagSerializer
+    pagination_class = None
 
 
 class FollowViewSet(mixins.ListModelMixin, mixins.CreateModelMixin,
                     viewsets.GenericViewSet):
     # НИГДЕ НЕ ИСПОЛЬЗУЕТСЯ
     pagination_class = LimitOffsetPagination
-    serializer_class = FollowSerializer
+    # serializer_class = FollowSerializer
     permission_classes = (IsAuthenticated, )
     filter_backends = (filters.SearchFilter, )
     search_fields = ('following__username',)
@@ -254,18 +263,26 @@ class FollowViewSet(mixins.ListModelMixin, mixins.CreateModelMixin,
 #         serializer.is_valid(raise_exception=True)
 #         serializer.save()
 #         return Response(serializer.data)
-    
-
-
-
-
-
 # class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
 #     queryset = IngredientIndividual.objects.all()
 #     serializer_class = IngredientIndividualSerializer
+
 
 class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
     queryset = Ingredient.objects.all()
     serializer_class = IngredientSerilizer
     filter_backends = (filters.SearchFilter, )
-    search_fields = ('name', ) 
+    # не работает
+    search_fields = ('name', )
+    pagination_class = None
+
+    # def get_queryset(self):
+    #     queryset = Ingredient.objects.all()
+    #     name = self.request.query_params.get('name', None)
+    #     if name is not None:
+    #         #  через ORM отфильтровать объекты модели Cat
+    #         #  по значению параметра color, полученного в запросе
+    #         queryset = queryset.filter(name=name)
+    #     return queryset
+
+
